IGNITE-23947: Sql. Add a test to check operator signatures
Fix between.
